refactor: extract tenant api dependency factories
